requirements.txt
python-telegram-bot==11.1.0
apiai==1.2.3
requests==2.21.0
gunicorn==19.9.0